// src/types/types.d.ts
interface AccountInfo extends WithId<Document> {
  _id?: ObjectId;
  firebaseId: string;
  name: string;
  address: string;
  grade: number;
  birthday: Date;
  email: string;
  tags: string[];
  checkIns: string[];
}
// ...
interface UserResponse {
  questionId: string;
  response: any;
}